Create AuthenticateUser.json
This JSON code defines a Postman request that tests an Azure Function called "AuthenticateUser". The request is defined as an item in a Postman collection and contains the following fields:

"info": contains metadata about the Postman collection, such as the name, the Postman ID, and the schema version.
"item": an array of requests and sub-collections that are part of the collection. In this case, it contains a single request object.
"name": the name of the request, which is "New Request".
"event": an array of events associated with the request, in this case, a single "test" event that is executed after the request has been sent.
"script": the JavaScript code that is executed during the "test" event. It contains two tests: one to check if the HTTP response status code is 200, and another to check if the message in the response JSON data is "Authenticated successfully".
"request": contains information about the HTTP request that is sent to the Azure Function. It specifies the HTTP method (POST), the URL of the Azure Function, and the request body, which contains the username and password used for authentication. The URL also includes a query parameter called "code" that is used to authenticate the request to the Azure Function.
"response": an empty array, which will be populated with the response data when the request is sent.